Minor changes to Arkansas

+++ catscripts/Government/Countries/United_States/US_states/Arkansas/Arkansas.sh
@@ -2,15 +2,18 @@
 
 KEYWORDS_ARKANSAS="Arkansas"
 
-ARKANSAS=`egrep -i "$KEYWORDS_ARKANSAS" newpages.txt > Arkansas.txt`
-
-
-if [ "$ARKANSAS" != "" ];
+if [ "$1" == "" ]; #Normal operation
 then
-  egrep -i "$KEYWORDS_ARKANSAS" newpages.txt > Arkansas.txt
-  export CATFILE="Arkansas.txt"
-  export CATNAME="Arkansas"
-  $CATEGORIZE
-  rm Arkansas.txt
-  unset ARKANSAS
+
+  ARKANSAS=`egrep -i "$KEYWORDS_ARKANSAS" newpages.txt`
+
+  if [ "$ARKANSAS" != "" ];
+  then
+    egrep -i "$KEYWORDS_ARKANSAS" newpages.txt > Arkansas.txt
+    export CATFILE="Arkansas.txt"
+    export CATNAME="Arkansas"
+    $CATEGORIZE
+    rm Arkansas.txt
+    unset ARKANSAS
+  fi
 fi
